fix: Error when storing JSON to DocumentDB that use dots in field names (#197)
To work around this issue, dots in field names will now get converted to backtick (`) characters when being stored, and converted back to dots upon retrieving. This has the consequence of making backticks reserved characters for field names, since if someone tries using one then that data will get incorrectly unescaped on retrieval. To prevent this situation, an error will now get returned if there's an attempt to store JSON data that uses backticks in field names.

Although this is a limitation of DocumentDB and not MongoDB, since this MongoDB storage provider can't reliably tell whether DocumentDB is being used or not, this dot replacement behaviour will also apply even if MongoDB is being used.

Signed-off-by: Derek Trider <Derek.Trider@securekey.com>