fix: Return error when using duplicate tag names in MongoDB and CouchDB
Put operations in the MongoDB and CouchDB operations that used the same tag name multiple times would result in all but the last usage getting lost due to the slice to map conversion that happens. With some effort, this issue may be fixable in the future, but for now I've updated the methods to return an explicit error to prevent unexpected behaviour.

Also updated golang.org/x/sys version in the mongodb module to resolve an issue when compiling with Go 1.18.

Also removed a TODO marker from the CouchDB implementation that is no longer needed.

Signed-off-by: Derek Trider <Derek.Trider@securekey.com>