feat: Direct JSON storage and custom indexing and querying for MongoDB
Added new methods to the MongoDB storage provider to allow a caller to make lower-level calls to MongoDB. These methods allow for more advanced indexing and querying, as well as making more efficient use of MongoDB possible. Note that this functionality will require making type assertions as they do not conform to the aries-framework-go interface.

Signed-off-by: Derek Trider <Derek.Trider@securekey.com>