feat: Changes to MongoDB custom method parameters
Changed the MongoDB CustomQuery method's filter parameter to be an interface{} to match the type used by the underlying driver. This allows the caller more flexibility in their queries.

Changed the MongoDB CustomQuery method's return parameter to be the specific MongoDB Iterator object instead of the storage interface's iterator type so that the caller can directly use the ValueAsRawMap method without doing a type assertion.

Signed-off-by: Derek Trider <Derek.Trider@securekey.com>